# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skriptra API
  version: 1.0.0
  description: |
    Exam intelligence platform API.

    **This contract is frozen.** The web client and the Go server are both built
    against it, and TypeScript types are generated from it, they are never
    hand-written twice. Breaking changes go to `/api/v2`, they do not mutate v1.

    ## Conventions

    - All paths are under `/api/v1`.
    - All IDs are UUIDs.
    - All timestamps are RFC 3339 UTC.
    - List endpoints are paginated and return a `PageMeta`.
    - Errors always use the `Error` envelope; the HTTP status carries the class,
      the `code` field carries the specific reason.
    - Authorization is enforced server-side on every request. The client renders
      permissions, it never decides them.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: /api/v1
    description: Same-origin deployment
...
security:
  - bearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: Register an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password:
                  type: string
                  minLength: 10
                  description: |
                    Length only. Composition rules push people towards
                    Password1! and are no longer recommended by NIST or OWASP.
                displayName: { type: string, maxLength: 200 }
      responses:
        "200":
          description: |
            Registered and signed in. The access token is in the body; the
            refresh token is set as an HttpOnly cookie and is never readable
            by client script.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: "skriptra_refresh, HttpOnly, SameSite=Lax, Secure in production."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "409":
          description: That email is already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /auth/login:
    post:
      tags: [auth]
      summary: Sign in
      description: |
        A wrong password and an unregistered address return the same `401` with
        the same message, and take the same time. Distinguishing them would
        turn this endpoint into an oracle for which addresses have accounts.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange the refresh cookie for a new access token
      description: |
        The refresh token is rotated on every use: the presented one is revoked
        and a new one issued. A stolen copy therefore stops working as soon as
        the legitimate client refreshes.

        Requires the cookie, so callers must send credentials.
      security: []
      responses:
        "200":
          description: New access token, and a rotated refresh cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401":
          description: No session, or it has expired or been revoked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /me:
    get:
      tags: [system]
      summary: Current user
      responses:
        "200":
          description: The authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /courses:
    get:
      tags: [courses]
      summary: List the current user's courses
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Course" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [courses]
      summary: Create a course
      description: |
        The caller becomes the course owner. Membership is what grants access:
        a course you do not belong to answers `404` on every endpoint, not
        `403`, since a `403` would confirm the identifier is real.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 200, examples: ["Linear Models"] }
                code: { type: string, maxLength: 50, examples: ["STAT-412"] }
                institution: { type: string, maxLength: 200, examples: ["TU Dortmund"] }
                language: { type: string, enum: [en, de], default: en }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /courses/{courseId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [courses]
      summary: Get a course with its rollup counts
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CourseDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /courses/{courseId}/chapters:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [chapters]
      summary: The course chapter taxonomy
      description: |
        Chapters are the primary navigation of the product and the filter that
        makes "give me the Chapter 3 questions" answerable. They are derived at
        course setup from a syllabus or textbook table of contents.
      responses:
        "200":
          description: Chapters in syllabus order
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Chapter" }
    post:
      tags: [chapters]
      summary: Save a confirmed taxonomy and sort existing questions into it
      description: |
        Replaces the taxonomy, then classifies every question in the course
        that has no chapter yet.

        The back-fill is the point: students upload the papers they have and
        find the syllabus afterwards, so without it a taxonomy would only ever
        apply to future uploads and everything already indexed would stay
        unclassified.

        A chapter removed from the taxonomy has its questions unclassified
        rather than left pointing at material the course no longer claims to
        cover. Chapters that survive keep their identifier, so an edit to a
        title does not discard existing assignments.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapters]
              properties:
                chapters:
                  type: array
                  minItems: 1
                  maxItems: 40
                  items: { $ref: "#/components/schemas/ProposedChapter" }
      responses:
        "200":
          description: The saved taxonomy, and how many questions it placed
          content:
            application/json:
              schema:
                type: object
                required: [data, questionsClassified]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Chapter" }
                  questionsClassified:
                    type: integer
                    description: |
                      How many previously unclassified questions the taxonomy
                      matched. The one number that tells the owner whether their
                      topic lists are any good.
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /courses/{courseId}/chapters/extract:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    post:
      tags: [chapters]
      summary: Propose a chapter taxonomy from a syllabus or contents page
      description: |
        Reads a chapter list out of an uploaded document and returns it without
        saving. Only the first pages are read, since a contents page is at the
        front by definition and rendering a whole textbook to find it is enough
        memory to kill the document service.

        Rules are tried first, which handles a conventional contents page
        exactly and for nothing. The model is the fallback for a syllabus
        written as prose.

        The result is a proposal because a contents page is messy: rules find
        most chapters and miss some, and only the course owner knows whether a
        line is a chapter or a heading inside one.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: A proposed taxonomy, not yet saved
          content:
            application/json:
              schema:
                type: object
                required: [chapters, source]
                properties:
                  chapters:
                    type: array
                    items: { $ref: "#/components/schemas/ProposedChapter" }
                  source:
                    type: string
                    enum: [rules, llm]
                    description: How the list was read, so a bad taxonomy can be traced to the method.
                  filename: { type: string }
        "415": { $ref: "#/components/responses/UnsupportedMedia" }
        "422":
          description: No chapter list could be read from that file
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /courses/{courseId}/exams:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [exams]
      summary: List exams, newest first
      parameters:
        - name: year
          in: query
          schema: { type: integer, minimum: 1990, maximum: 2100 }
        - name: term
          in: query
          schema: { $ref: "#/components/schemas/Term" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Exams
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Exam" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
  /courses/{courseId}/questions:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [questions]
      summary: List questions with filters
      description: |
        This is the `enumerate` path, exhaustive, ordered and paginated.
        It is deliberately **not** vector search: a request for every Chapter 3
        question must return every Chapter 3 question, which top-k retrieval
        structurally cannot guarantee.
      parameters:
        - name: chapterId
          in: query
          schema: { type: string, format: uuid }
        - name: chapterNumber
          in: query
          description: Convenience filter so the UI can link by number.
          schema: { type: integer, minimum: 1 }
        - name: yearFrom
          in: query
          schema: { type: integer }
        - name: yearTo
          in: query
          schema: { type: integer }
        - name: term
          in: query
          schema: { $ref: "#/components/schemas/Term" }
        - name: q
          in: query
          description: Optional full-text filter applied on top of the structured filters.
          schema: { type: string }
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, chapter]
            default: newest
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Question" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
  /courses/{courseId}/documents:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [documents]
      summary: List documents in a course
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/IngestStatus" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Document" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [documents]
      summary: Upload a document
      description: |
        Accepts the file and returns immediately with status `queued`; ingestion
        runs asynchronously via NATS. Poll `/documents/{id}/status` for progress.

        Files are content-hashed on upload. Re-uploading a document that already
        exists in the course returns `200` with the existing document rather than
        creating a duplicate.

        Deduplication applies only to a document that is usable. A previous
        attempt that **failed**, or one abandoned mid-ingestion for more than
        thirty minutes, is re-queued and returns `202`: the row is written
        before ingestion runs, so treating a failure as a duplicate would make
        it permanent and leave the user re-uploading a file they are told
        already exists.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, kind]
              properties:
                file:
                  type: string
                  format: binary
                kind: { $ref: "#/components/schemas/DocumentKind" }
                year: { type: integer }
                term: { $ref: "#/components/schemas/Term" }
      responses:
        "200":
          description: Document already existed (deduplicated by content hash)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "202":
          description: Accepted and queued for ingestion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMedia" }
  /courses/{courseId}/analytics/chapter-frequency:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [analytics]
      summary: How often each chapter is tested
      description: |
        A pure SQL aggregate: the `analyse` intent never touches the LLM.
        This is what makes Skriptra more than a chatbot.
      parameters:
        - name: yearFrom
          in: query
          schema: { type: integer }
        - name: yearTo
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Frequency per chapter, syllabus order
          content:
            application/json:
              schema:
                type: object
                required: [data, totalQuestions]
                properties:
                  totalQuestions: { type: integer }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ChapterFrequency" }
  /courses/{courseId}/conversations:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    get:
      tags: [conversations]
      summary: List the current user's conversations in a course
      description: |
        Most recently active first. A conversation is titled from the first
        question asked in it, which is what a student recognises it by.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Conversation" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
  /conversations/{conversationId}:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
    get:
      tags: [conversations]
      summary: Get a conversation with every message in it
      description: |
        Messages are oldest first. Citations are stored on the message rather
        than resolved at read time, so an answer reads exactly as it was given
        even after the document it cited has been deleted or re-ingested.

        A conversation belonging to another user returns `404`, not `403`: a
        `403` would confirm the identifier is real.
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [conversations]
      summary: Delete a conversation and its messages
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------------ exams --
  /exams/{examId}:
    parameters:
      - name: examId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [exams]
      summary: Get an exam with its questions
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExamDetail" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------------------------------------- questions --
  /questions/{questionId}:
    parameters:
      - $ref: "#/components/parameters/QuestionId"
    get:
      tags: [questions]
      summary: Get a single question
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /questions/{questionId}/similar:
    parameters:
      - $ref: "#/components/parameters/QuestionId"
    get:
      tags: [questions]
      summary: Find similar questions across years
      description: |
        The `similar` intent, vector k-NN over question embeddings, excluding
        the question itself. This is the feature that turns a pile of PDFs into
        a study tool.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
        - name: minScore
          in: query
          schema: { type: number, minimum: 0, maximum: 1, default: 0.5 }
      responses:
        "200":
          description: Similar questions, most similar first
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SimilarQuestion" }

  # ----------------------------------------------------------------- search --
  /search:
    post:
      tags: [search]
      summary: Hybrid search over a course corpus
      description: |
        Dense (pgvector cosine) + sparse (Postgres full-text) fused with
        reciprocal rank fusion, with structured filters applied as WHERE clauses
        rather than post-filtering. Returns passages, not an answer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseId, query]
              properties:
                courseId: { type: string, format: uuid }
                query: { type: string, minLength: 1 }
                filters: { $ref: "#/components/schemas/RetrievalFilters" }
                limit: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        "200":
          description: Ranked passages
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SearchHit" }

  # -------------------------------------------------------------------- ask --
  /ask:
    post:
      tags: [ask]
      summary: Ask a natural-language question about a course
      description: |
        Routes the query by intent before doing any work:

        | Intent | Path |
        |---|---|
        | `enumerate` | SQL over `questions`, exhaustive |
        | `explain`   | Hybrid retrieval, then the LLM, with citations |
        | `similar`   | Vector k-NN |
        | `analyse`   | SQL aggregate, no LLM |
        | `hybrid`    | enumerate to select, then explain |

        Set `stream: true` to receive `text/event-stream` instead of JSON.

        **SSE event sequence:** `intent` (once, immediately, lets the UI show
        what it decided) → `sources` (once, before generation, so citations can
        render while tokens stream) → `token` (many) → `done` (once, with usage)
        or `error`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseId, question]
              properties:
                courseId: { type: string, format: uuid }
                question: { type: string, minLength: 1, maxLength: 4000 }
                conversationId:
                  type: string
                  format: uuid
                  description: Omit to start a new conversation.
                filters: { $ref: "#/components/schemas/RetrievalFilters" }
                stream: { type: boolean, default: false }
      responses:
        "200":
          description: |
            An answer. `application/json` when `stream` is false,
            `text/event-stream` when true.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Answer" }
            text/event-stream:
              schema:
                type: string
                description: |
                  Newline-delimited SSE frames.

                  ```
                  event: intent
                  data: {"intent":"explain"}

                  event: sources
                  data: {"sources":[{...}]}

                  event: token
                  data: {"text":"Maximum "}

                  event: done
                  data: {"conversationId":"...","messageId":"...","usage":{...}}
                  ```
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "503":
          description: The configured model provider is unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # -------------------------------------------------------------- documents --
  /documents/{documentId}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [documents]
      summary: Get document metadata
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [documents]
      summary: Delete a document and everything derived from it
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Session:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: |
            Short-lived JWT, sent as `Authorization: Bearer`. Held in memory by
            the client, never in localStorage, where any script on the page
            could read it.
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer, description: Seconds until the access token expires. }
        user: { $ref: "#/components/schemas/User" }
...